+++ crates/miden-testing/src/kernel_tests/tx/test_account.rs
@@ -91,13 +91,12 @@ pub fn test_account_type() -> miette::Result<()> {
 
             let code = format!(
                 "
-                use.kernel::account
+                use.kernel::account_id
 
                 begin
-                    exec.account::{}
+                    exec.account_id::{procedure}
                 end
-                ",
-                procedure
+                "
             );
 
             let process = CodeExecutor::with_advice_provider(MemAdviceProvider::default())
@@ -119,7 +118,7 @@ pub fn test_account_type() -> miette::Result<()> {
             );
         }
 
-        assert!(has_type, "missing test for type {:?}", expected_type);
+        assert!(has_type, "missing test for type {expected_type:?}");
     }
 
     Ok(())
@@ -161,10 +160,10 @@ pub fn test_account_validate_id() -> miette::Result<()> {
         let suffix = Felt::try_from((account_id % (1u128 << 64)) as u64).unwrap();
 
         let code = "
-            use.kernel::account
+            use.kernel::account_id
 
             begin
-                exec.account::validate_id
+                exec.account_id::validate
             end
             ";
 
@@ -213,11 +212,11 @@ fn test_is_faucet_procedure() -> miette::Result<()> {
 
         let code = format!(
             "
-            use.kernel::account
+            use.kernel::account_id
 
             begin
                 push.{prefix}
-                exec.account::is_faucet
+                exec.account_id::is_faucet
                 # => [is_faucet, account_id_prefix]
 
                 # truncate the stack
@@ -235,8 +234,7 @@ fn test_is_faucet_procedure() -> miette::Result<()> {
         assert_eq!(
             process.stack.get(0),
             Felt::new(is_faucet as u64),
-            "Rust and Masm is_faucet diverged. account_id: {}",
-            account_id
+            "Rust and MASM is_faucet diverged for account_id {account_id}"
         );
     }
 
@@ -647,7 +645,7 @@ fn test_account_component_storage_offset() -> miette::Result<()> {
     "
     );
     let tx_script_program = assembler.assemble_program(tx_script_source_code).unwrap();
-    let tx_script = TransactionScript::new(tx_script_program, vec![]);
+    let tx_script = TransactionScript::new(tx_script_program);
 
     // setup transaction context
     let tx_context = TransactionContextBuilder::new(account.clone()).tx_script(tx_script).build();